HW3
Write a program that processes input from a user about a playing card.  The program should ask the user for the suit of the card (hearts, diamonds, spades, clubs).  If the user enters hearts or diamonds, the program should say "That card is Red".  If the user enters spades or clubs, the program should say "That card is Black".  The program should ask if for the face value of the card (1-13; 1 for Ace, etc.).  Assume that you are playing BlackJack.