Inclass 4
For this assignment, you want to create a Python program that prompts the user for some input and writes to an output file that is formatted as XML.  XML (Extensible Markup Language) is a markup language that is better suited for data storage than HTML. The program will ask the user to enter an animal or 'done' when done.  The animals entered will be written to an XML file.