Inclass 6
You are to write a Python program that will ask for the name of the file to process and write out the data in XML format.  On each line of the input text file, you have four values.  The first value is the subject's name, the second value is the subject's sex, the third value is thesubject's height in inches and the fourth value is the subject's weight in pounds.